:bug: allow approving a skill
